Split out AttachFile to its own control
